chore: move email service files to common app
