+++ bc_obps/service/application_access_service.py
@@ -1,6 +1,6 @@
 from uuid import UUID
-from common_utils.email.email_service import EmailService
-from common_utils.enums import AdminAccessRequestStates
+from common.service.email.email_service import EmailService
+from common.enums import AdminAccessRequestStates
 from registration.models import UserOperator
 from service.data_access_service.operator_service import OperatorDataAccessService
 from service.data_access_service.user_operator_service import UserOperatorDataAccessService
